requirements.txt
google-api-python-client==2.118.0
google-auth==2.28.0
Pillow==10.2.0
boto3==1.34.34
loguru==0.7.2
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
